完善陷阱 完善数值伤害

+++ Assets/Scripts/GameEntity/Data/HurtData.cs
@@ -5,7 +5,7 @@ using System;
 public class HurtData : BaseEventArgs
 {
     public Vector2 AttackForce { get; set; }
-    public float AttackValue { get; set; }
+    public int AttackValue { get; set; }
     public int AttackerID { get; set; }
     public bool IsSwoon { get; set; }//是否击飞
     public float AttackerDir { get; set; }
